requirements.txt
backports.functools-lru-cache==1.5
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
docutils==0.14
idna==2.7
Kivy==1.10.0
Kivy-Garden==0.1.4
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.1.0
pygame==1.9.3
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
scipy==1.1.0
six==1.11.0
subprocess32==3.5.2
torch==0.4.0
torchvision==0.2.1
urllib3==1.23
